# Lammah-game-backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/ai-generator/debug-tools": {
      "get": {
        "operationId": "aiDebugTools",
        "summary": "Debug local media tool availability for AI asset generation",
        "deprecated": true,
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiToolDiagnosticsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/ai-generator/generate-reviewed": {
      "post": {
        "operationId": "aiGenerateReviewed",
        "summary": "Generate reviewed AI question drafts without saving them",
        "description": "Generates multiple-choice question drafts from the configured AI provider, validates quality issues, and returns drafts only.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReviewedQuestionsDto"
              },
              "examples": {
                "default": {
                  "summary": "Generate reviewed Arabic World Cup questions",
                  "value": {
                    "catalogName": "رياضة",
                    "categoryName": "كأس العالم",
                    "difficulty": "medium",
                    "count": 10,
                    "language": "ar"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reviewed question drafts generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateReviewedQuestionsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No reviewed drafts were approved; bounded source and candidate diagnostics are returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateReviewedQuestionsErrorResponseDto"
                }
              }
            }
          },
          "503": {
            "description": "AI question generation is disabled",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 503,
                    "code": "AI_QUESTION_GENERATION_DISABLED",
                    "message": "AI question generation is currently disabled."
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/admin/ai-generator/save-drafts": {
      "post": {
        "operationId": "aiSaveReviewedDrafts",
        "summary": "Persist reviewed questions as unpublished AI drafts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveReviewedDraftsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveReviewedDraftsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin AI Generator"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Lammah Game API",
    "description": "Runtime HTTP contract for Lammah game administration, content, and gameplay.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "QuestionGameplayType": {
        "type": "string",
        "default": "standard",
        "enum": [
          "standard",
          "ranked_list"
        ]
      },
      "LocalizedQuestionTextDto": {
        "type": "object",
        "properties": {
          "ar": {
            "type": "string"
          },
          "en": {
            "type": "string"
          }
        },
        "required": [
          "ar"
        ]
      },
...
      "RankedListDefinitionDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "$ref": "#/components/schemas/LocalizedQuestionTextDto"
          },
          "entries": {
            "minItems": 10,
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RankedListEntryDto"
            }
          }
        },
        "required": [
          "displayName",
          "entries"
        ]
      },
      "DifficultyLevel": {
        "type": "string",
        "enum": [
          "easy",
          "medium",
          "hard"
        ]
      },
...
      "GameMode": {
        "type": "string",
        "enum": [
          "trivia",
          "identifyCharacter",
          "identifyVoice",
          "identifyImage",
          "completeQuote",
          "timeline",
          "emojiPuzzle",
          "identifySong",
          "identifySinger",
          "identifyMusicIntro"
        ]
      },
...
      "AudioQuestionKind": {
        "type": "string",
        "enum": [
          "identify_song",
          "identify_artist",
          "identify_character",
          "identify_voice",
          "identify_game",
          "identify_movie",
          "identify_dialogue_source",
          "identify_sound_effect",
          "custom"
        ]
      },
      "QuestionAudioRequestDto": {
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/AudioQuestionKind"
          },
          "searchQuery": {
            "type": "string"
          },
          "targetName": {
            "type": "string"
          },
          "sourceTitle": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "preferredStartSeconds": {
            "type": "number",
            "minimum": 0
          },
          "preferredDurationSeconds": {
            "type": "number",
            "minimum": 3,
            "maximum": 20
          },
          "provider": {
            "type": "string"
          },
          "requestVersion": {
            "type": "number",
            "minimum": 1,
            "readOnly": true
          },
          "requestHash": {
            "type": "string",
            "readOnly": true
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "selectedCandidateId": {
            "type": "string",
            "nullable": true,
            "readOnly": true
          },
          "candidateSetVersion": {
            "type": "integer",
            "minimum": 1,
            "nullable": true,
            "readOnly": true
          }
        },
        "required": [
          "kind",
          "searchQuery"
        ]
      },
...
      "QuestionStatus": {
        "type": "string",
        "enum": [
          "draft",
          "approved",
          "published",
          "archived",
          "rejected"
        ]
      },
      "QuestionSource": {
        "type": "string",
        "enum": [
          "manual",
          "ai",
          "imported",
          "music"
        ]
      },
...
      "ResolvedQuestionMediaResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "image",
              "audio",
              "video"
            ]
          },
          "url": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          }
        },
        "required": [
          "type",
          "url"
        ]
      },
...
      "QuestionAssetResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/QuestionAssetType"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sourceUrl": {
            "type": "string"
          },
          "searchQuery": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "duplicateDiagnostics": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "type",
          "url",
          "source"
        ]
      },
...
      "QuestionResponseDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "question": {
            "type": "string"
          },
          "questionType": {
            "$ref": "#/components/schemas/QuestionGameplayType"
          },
          "text": {
            "$ref": "#/components/schemas/LocalizedQuestionTextDto"
          },
          "maxPoints": {
            "type": "number"
          },
          "turnDurationSeconds": {
            "type": "number"
          },
          "maxStrikesPerTeam": {
            "type": "number"
          },
          "rankedList": {
            "$ref": "#/components/schemas/RankedListDefinitionDto"
          },
          "answer": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "wrongAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "acceptedAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "explanation": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyLevel"
          },
          "points": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "score": {
            "type": "number",
            "enum": [
              200,
              400,
              600
            ]
          },
          "gameMode": {
            "$ref": "#/components/schemas/GameMode"
          },
          "type": {
            "type": "string"
          },
          "preferredPresentationType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video",
              "gif"
            ]
          },
          "effectivePresentationType": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "video"
            ]
          },
          "mediaAvailable": {
            "type": "boolean"
          },
          "mediaFallbackReason": {
            "type": "string",
            "enum": [
              "NO_MEDIA",
              "NOT_READY",
              "PROCESSING",
              "FAILED",
              "REJECTED",
              "STALE",
              "MISSING_ASSET",
              "INVALID_ASSET"
            ],
            "nullable": true
          },
          "resolvedMedia": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ResolvedQuestionMediaResponseDto"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "source": {
            "$ref": "#/components/schemas/QuestionSource"
          },
          "primaryAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "requiresAudio": {
            "type": "boolean",
            "default": false
          },
          "audioKind": {
            "$ref": "#/components/schemas/AudioQuestionKind"
          },
          "audioRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAudioRequestDto"
              }
            ]
          },
          "audioCandidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionAudioCandidateResponseDto"
            }
          },
          "audioStatus": {
            "$ref": "#/components/schemas/AudioAssetStatus"
          },
          "audioAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "audioReviewStatus": {
            "$ref": "#/components/schemas/AudioReviewStatus"
          },
          "audioDiagnostics": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "audioRequestStale": {
            "type": "boolean"
          },
          "coverImage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionAssetResponseDto"
              }
            ]
          },
          "primaryAssetRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageRequest": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true
          },
          "coverImageStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "coverImageFailureReason": {
            "type": "string"
          },
          "assetStatus": {
            "$ref": "#/components/schemas/AssetStatus"
          },
          "assetFailureReason": {
            "type": "string"
          },
          "assetFailureStep": {
            "type": "string"
          },
          "assetFailureDiagnostics": {
            "type": "object",
            "additionalProperties": true
          },
          "gameplayMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "aiMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "mediaUrl": {
            "type": "string"
          },
          "hasPreviewAudio": {
            "type": "boolean"
          },
          "isFreeGameQuestion": {
            "type": "boolean"
          },
          "qualityScore": {
            "type": "number"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "_id",
          "question",
          "questionType",
          "wrongAnswers",
          "difficulty",
          "status",
          "source",
          "requiresAudio",
          "audioStatus"
        ]
      },
...
      "AiToolDiagnosticsResponseDto": {
        "type": "object",
        "properties": {
          "ffmpegAvailable": {
            "type": "boolean"
          },
          "ytDlpAvailable": {
            "type": "boolean"
          },
          "ffmpegVersion": {
            "type": "string"
          },
          "ytDlpVersion": {
            "type": "string"
          }
        },
        "required": [
          "ffmpegAvailable",
          "ytDlpAvailable",
          "ffmpegVersion",
          "ytDlpVersion"
        ]
      },
      "GenerateReviewedQuestionsDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string"
          },
          "catalogName": {
            "type": "string"
          },
          "categoryName": {
            "type": "string"
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "count": {
            "type": "number",
            "minimum": 1,
            "maximum": 20,
            "default": 2
          },
          "language": {
            "type": "string",
            "enum": [
              "ar"
            ]
          },
          "strategy": {
            "type": "string",
            "enum": [
              "source-curated"
            ],
            "default": "source-curated"
          },
          "sourceIds": {
            "example": [
              "open-trivia-db"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "allowGeneratedFallback": {
            "type": "boolean",
            "default": false,
            "description": "Reserved for future use. Generated fallback is currently disabled."
          }
        }
      },
...
      "ReviewedQuestionDraftResponseDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "correctAnswer": {
            "type": "string"
          },
          "wrongAnswers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "medium",
              "hard"
            ]
          },
          "gameMode": {
            "type": "string",
            "enum": [
              "trivia",
              "identifyCharacter",
              "identifyVoice",
              "identifyImage",
              "completeQuote",
              "timeline",
              "emojiPuzzle",
              "identifySong",
              "identifySinger",
              "identifyMusicIntro"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "text",
              "image",
              "audio",
              "quote",
              "emoji",
              "timeline",
              "video",
              "gif"
            ]
          },
          "assetRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "assetStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "asset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "primaryAssetRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "primaryAssetStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "primaryAsset": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "coverImageRequest": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetRequestResponseDto"
              }
            ]
          },
          "coverImageStatus": {
            "type": "string",
            "enum": [
              "NOT_REQUIRED",
              "PENDING",
              "READY",
              "FAILED"
            ]
          },
          "coverImage": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/AiAssetResponseDto"
              }
            ]
          },
          "coverImageFailureReason": {
            "type": "string",
            "nullable": true
          },
          "assetFailureReason": {
            "type": "string"
          },
          "assetFailureStep": {
            "type": "string"
          },
          "assetFailureDiagnostics": {
            "oneOf": [
              {
                "type": "object",
                "additionalProperties": true
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            ]
          },
          "wasGameplayAutoFixed": {
            "type": "boolean"
          },
          "gameplayFixReason": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "qualityScore": {
            "type": "number"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "agentTrace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AiAgentTraceResponseDto"
            }
          },
          "gameplayMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "aiMetadata": {
            "type": "object",
            "additionalProperties": true
          },
          "verificationDiagnostics": {
            "$ref": "#/components/schemas/AiVerificationDiagnosticsResponseDto"
          }
        },
        "required": [
          "question",
          "correctAnswer",
          "wrongAnswers",
          "difficulty",
          "gameMode",
          "type",
          "assetStatus",
          "primaryAssetStatus",
          "coverImageStatus",
          "explanation",
          "qualityScore",
          "issues"
        ]
      },
      "ReviewedQuestionsDataResponseDto": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewedQuestionDraftResponseDto"
            }
          }
        },
        "required": [
          "questions"
        ]
      },
      "GenerateReviewedQuestionsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 201
          },
          "message": {
            "type": "string"
          },
          "count": {
            "type": "number"
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          },
          "data": {
            "$ref": "#/components/schemas/ReviewedQuestionsDataResponseDto"
          }
        },
        "required": [
          "statusCode",
          "message",
          "count",
          "meta",
          "data"
        ]
      },
      "CandidateValidationResultResponseDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "PASS",
              "FAIL",
              "NOT_EVALUATED"
            ]
          },
          "issueCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "status",
          "issueCodes"
        ]
      },
      "SourceCandidateDiagnosticResponseDto": {
        "type": "object",
        "properties": {
          "sourceId": {
            "type": "string"
          },
          "sourceQuestionId": {
            "type": "string"
          },
          "sourceQuestion": {
            "type": "string"
          },
          "sourceAnswer": {
            "type": "string"
          },
          "curatedQuestion": {
            "type": "string",
            "nullable": true
          },
          "curatedAnswer": {
            "type": "string",
            "nullable": true
          },
          "semanticFingerprint": {
            "type": "string"
          },
          "duplicateScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "validationResult": {
            "$ref": "#/components/schemas/CandidateValidationResultResponseDto"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "CREATED",
              "REJECTED",
              "FAILED",
              "NOT_SELECTED"
            ]
          },
          "rejectionReason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "sourceId",
          "sourceQuestionId",
          "sourceQuestion",
          "sourceAnswer",
          "curatedQuestion",
          "curatedAnswer",
          "semanticFingerprint",
          "duplicateScore",
          "validationResult",
          "outcome",
          "rejectionReason"
        ]
      },
      "GenerateReviewedQuestionsErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "AI pipeline produced no drafts"
          },
          "issueCodes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          },
          "sourceDiagnostics": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "sourceSummary": {
            "type": "object",
            "additionalProperties": true
          },
          "candidateDiagnostics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SourceCandidateDiagnosticResponseDto"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "issueCodes",
          "meta",
          "sourceDiagnostics",
          "sourceSummary",
          "candidateDiagnostics"
        ]
      },
      "SaveReviewedDraftsDto": {
        "type": "object",
        "properties": {
          "drafts": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          },
          "categoryId": {
            "type": "string"
          },
          "catalogId": {
            "type": "string"
          }
        },
        "required": [
          "drafts",
          "categoryId"
        ]
      },
      "SaveDraftFailureDto": {
        "type": "object",
        "properties": {
          "index": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "index",
          "reason"
        ]
      },
      "SaveReviewedDraftsResponseDto": {
        "type": "object",
        "properties": {
          "savedCount": {
            "type": "number",
            "minimum": 0
          },
          "failedCount": {
            "type": "number",
            "minimum": 0
          },
          "savedQuestions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionResponseDto"
            }
          },
          "failures": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaveDraftFailureDto"
            }
          }
        },
        "required": [
          "savedCount",
          "failedCount",
          "savedQuestions",
          "failures"
        ]
      },
...
    }
  }
}